Add files via upload
Geckodriver works as webdriver of Mozilla browser. So, whenever we want to automate website and we want to use Mozilla as browser instead of other browser. related code for automation is in this repository, Download and run them, they are quite intersting.
If you want to automate Chrome browser then Download ChromeDriver for that.